requirements.txt
flask~=3.0.2
requests~=2.31.0
torch~=2.2.2
torchvision~=0.17.2
pillow~=10.3.0